Windows: make remote control work over a tcp address
kitty @ and listen_on were documented as unavailable on Windows because kitty
drives them over a Unix socket and CPython does not build socket.AF_UNIX there.
The fact is right and unchanged: python/cpython#77589 has been open since 2018,
Python 3.14 still has no AF_UNIX, and passing family 1 by number gets as far as
a socket object before failing on "bind(): bad family". Windows itself has
carried AF_UNIX since build 17063, so this is CPython's gap, not the platform's.

The conclusion drawn from it was wrong. Remote control never needed AF_UNIX,
because listen_on has always taken a tcp: address as well. Two bugs stood in the
way of that working, one at each end.

listen_on() tested socket.AF_UNIX for every address family, not only for unix:,
so a tcp: address raised AttributeError before it could bind, and the caller
turned that into "Invalid listen_on, ignoring" with the reason discarded. The
test was redundant anyway: socket_path is set only for a non-abstract unix:
address, so it already implies the family.

With the listener up, connections were accepted and then answered with nothing,
which a client sees as an i/o timeout because TCP completes a handshake from the
backlog whether or not anything services it. kitty's own stderr carried the
answer, a bare MemoryError: peer_message_received is called with a y# format,
whose length is a Py_ssize_t, and the length was cast to int. On the SysV ABI
that argument still arrives in a register and writing its low half zeroes the
top, so this is invisible on Linux and macOS. Win64 passes four arguments in
registers, which leaves this one first on the stack, where four bytes of stale
stack became the top half of the length. Every command arrived with a nonsense
size and allocation failed.

Verified both ways: kitten @ ls from outside with --to, and from inside a kitty
window with no address at all, resolving KITTY_LISTEN_ON. Both return the
instance JSON, and stderr is clean.

Also, listen_on failures now carry their reason rather than all reading alike,
and a unix: address on a Python without the family says so and points at tcp:
instead of failing as a parse error.

One caveat is documented rather than fixed: a loopback port has no ACL, where a
Unix socket has file permissions and a peer uid check. remote_control_password
matters more here, and a named pipe is the shape a stricter transport would take.

+++ kitty/child-monitor.c
@@ -674,7 +674,15 @@ parse_input(ChildMonitor *self) {
             Message *msg = msgs + i;
             PyObject *resp = NULL;
             if (msg->data) {
-                resp = PyObject_CallMethod(global_state.boss, "peer_message_received", "y#KO", msg->data, (int)msg->sz, msg->peer_id, msg->is_remote_control_peer ? Py_True : Py_False);
+                // The length must be passed as Py_ssize_t: PY_SSIZE_T_CLEAN is
+                // set, so y# reads eight bytes here. Passing an int is harmless
+                // on the SysV ABI, where this argument still lands in a register
+                // and the write to its low half zeroes the top. Win64 passes only
+                // four arguments in registers, so this one is the first on the
+                // stack, and four bytes of stale stack were being read as the top
+                // half of the length. The result was a nonsense size and a
+                // MemoryError instead of a remote control command.
+                resp = PyObject_CallMethod(global_state.boss, "peer_message_received", "y#KO", msg->data, (Py_ssize_t)msg->sz, msg->peer_id, msg->is_remote_control_peer ? Py_True : Py_False);
                 free(msg->data);
                 if (!resp) PyErr_Print();
             }
